refactor: splits validator and DAOSignatureBlobChecker into two separate files
